split traits
