+++ poseidon-parameters/Cargo.toml
@@ -9,6 +9,7 @@ license = "MIT OR Apache-2.0"
 [dependencies]
 ark-ff = { version = "^0.3.0", default-features = false }
 ark-std = { version = "^0.3.0", default-features = false }
+num-integer = { version = "0.1.45", default-features = false }
 
 # [features]
 # default = ["std"]
