+++ poseidon-paramgen/src/matrix/mult.rs
@@ -1,6 +1,7 @@
 use anyhow::{anyhow, Result};
 use ark_ff::PrimeField;
 use ark_std::{ops::Mul, vec::Vec};
+use poseidon_parameters::BasicMatrixOperations;
 
 use crate::{Matrix, MatrixOperations, SquareMatrix};
 
@@ -14,7 +15,10 @@ pub fn dot_product<F: PrimeField>(a: &[F], b: &[F]) -> F {
 }
 
 /// Multiply two matrices
-pub fn mat_mul<F: PrimeField, M: MatrixOperations<F>>(lhs: &M, rhs: &M) -> Result<M> {
+pub fn mat_mul<F: PrimeField, M: BasicMatrixOperations<F> + MatrixOperations<F>>(
+    lhs: &M,
+    rhs: &M,
+) -> Result<M> {
     if lhs.n_cols() != rhs.n_rows() {
         return Err(anyhow!(
             "matrix dimensions do not allow matrix multiplication"
