+++ poseidon-paramgen/src/matrix/square.rs
@@ -1,6 +1,7 @@
 use anyhow::{anyhow, Result};
 use ark_ff::PrimeField;
 use ark_std::{vec, vec::Vec};
+use poseidon_parameters::BasicMatrixOperations;
 
 use crate::{mat_mul, Matrix, MatrixOperations, SquareMatrixOperations};
 
@@ -8,33 +9,33 @@ use crate::{mat_mul, Matrix, MatrixOperations, SquareMatrixOperations};
 pub struct SquareMatrix<T>(pub T);
 
 impl<F: PrimeField> MatrixOperations<F> for poseidon_parameters::SquareMatrix<F> {
-    fn new(n_rows: usize, n_cols: usize, elements: Vec<F>) -> Self {
-        Self(poseidon_parameters::Matrix::new(n_rows, n_cols, elements))
-    }
+    // fn new(n_rows: usize, n_cols: usize, elements: Vec<F>) -> Self {
+    //     Self(poseidon_parameters::Matrix::new(n_rows, n_cols, elements))
+    // }
 
-    fn elements(&self) -> &Vec<F> {
-        self.0.elements()
-    }
+    // fn elements(&self) -> &Vec<F> {
+    //     self.0.elements()
+    // }
 
-    fn n_rows(&self) -> usize {
-        self.0.n_rows
-    }
+    // fn n_rows(&self) -> usize {
+    //     self.0.n_rows
+    // }
 
-    fn n_cols(&self) -> usize {
-        self.0.n_cols
-    }
+    // fn n_cols(&self) -> usize {
+    //     self.0.n_cols
+    // }
 
-    fn get_element(&self, i: usize, j: usize) -> F {
-        self.0.get_element(i, j)
-    }
+    // fn get_element(&self, i: usize, j: usize) -> F {
+    //     self.0.get_element(i, j)
+    // }
 
-    fn set_element(&mut self, i: usize, j: usize, val: F) {
-        self.0.set_element(i, j, val)
-    }
+    // fn set_element(&mut self, i: usize, j: usize, val: F) {
+    //     self.0.set_element(i, j, val)
+    // }
 
-    fn rows(&self) -> Vec<&[F]> {
-        self.0.rows()
-    }
+    // fn rows(&self) -> Vec<&[F]> {
+    //     self.0.rows()
+    // }
 
     /// Take transpose of the matrix
     fn transpose(&self) -> Self {
